Move account forms into admin
To be in the same app the views are defined
