+++ wagtail/admin/views/account.py
@@ -14,12 +14,12 @@ from django.utils.translation import gettext as _
 from django.utils.translation import gettext_lazy, override
 from django.views.decorators.debug import sensitive_post_parameters
 
+from wagtail.admin.forms.account import (
+    AvatarPreferencesForm, LocalePreferencesForm, NameEmailForm, NotificationPreferencesForm)
 from wagtail.admin.forms.auth import LoginForm, PasswordChangeForm, PasswordResetForm
 from wagtail.admin.localization import get_available_admin_languages, get_available_admin_time_zones
 from wagtail.core import hooks
 from wagtail.core.models import UserPagePermissionsProxy
-from wagtail.users.forms import (
-    AvatarPreferencesForm, LocalePreferencesForm, NameEmailForm, NotificationPreferencesForm)
 from wagtail.users.models import UserProfile
 from wagtail.utils.loading import get_custom_form
